requirements.txt
youtube-transcript-api==1.2.0
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4>=4.12.3
openai>=1.40.0

